fix random_track_and_album bug, make OutputStream Arc
we now make cheap copies of the stream on when cloning  and recreate the sink on Player::stop. this ensures the sink is cleared when on Player::stop. previously behaviour was non deterministic as methods on Rodio::Sink are non blocking - in our case the first call to Player::random_track_and_album would also throw away our newly appended track