switch to crossterm backend (from ncurses)
optimise rendering by impl caching for player_view, reducing fps for
player_view, adding early exit for PlayerView::draw when finder is shown
(as the full view stack is drawn, not just top view), switching to
event-driven rendering in finder_view (setting fps to 0).

crossterm does not seem to be as performant as ncurses (at least on
macOS). these changes are to help mitigate screen-tearing in the
player_view